Building (and working!) sdk installer for Windows
Currently all this does is take a snapshot of the nacl_sdk directory under
the nacl tree, and package it into an .msi file.

The installer is basic--the only option is to set the install directory.

SCons is set to build the msi on Windows only. The target name is sdk_msi.


Review URL: http://codereview.chromium.org/303004

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@895 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2