gcc: Fix compilation of computed gotos
This brings gcc-4.2.2.patch into line with gcc-4.4.1.patch.

The "%A0" syntax adds a "*" in front of the register name which isn't
the right syntax for nacljmp.

Without this change, the compile will fail with an error like this:
/tmp/ccH6hCel.s: Assembler messages:
/tmp/ccH6hCel.s:14: Error: suffix or operands invalid for `nacljmp' 

This change also adds a test case for the toolchain.
The test can be run with, for example, the following command:
./scons MODE=opt-linux,nacl sdk_minimal_test

Review URL: http://codereview.chromium.org/348032

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@962 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2