Added back address space layout checks that disappeared due to a careless merge -- from the elf_util refactoring.  Also adds test data (hand edited binaries that no sane elf tool could generate) that tickles the checks.
Also added an explicit flag (for now) to sel_main to enable/disable
shared memory for dynamic text.

NB: the trybots are useless for the new tests, since the testdata are
nexes and binary data is not sent to the trybots, so none of the new
tests can run.

Review URL: http://codereview.chromium.org/464025

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@1075 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2