Add syscall wrapper for dynamic code loading
I am committing this before the syscall implementation because:
 * Syscall header changes require a lengthy newlib rebuild.
   The target extra_sdk_update(_header) doesn't cover syscall headers.
   (However, the shortcut is to do:
   cp src/trusted/service_runtime/include/sys/nacl_syscalls.h \
     toolchain/linux_x86/sdk/nacl-sdk/nacl64/include/sys/nacl_syscalls.h)
 * Let's get any bikeshedding about the syscall name out of the way first. :-)
 * Incremental is better!

I'm calling the syscall dyncode_copy because it's analogous to
memcpy() (but with a little extra logic).  In principle we might add
another syscall that loads code from a file descriptor; this could be
called dyncode_read or dyncode_map.  The "nacl_" prefix in the helper
function is to reduce the chance of naming collisions.

Review URL: http://codereview.chromium.org/1560019

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@2019 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2