Fix shared memory to work again in the Chromium Mac sandbox
This fixes breakage introduced by r2097, which stopped shared memory
creation from working in the context of Chromium's outer sandbox on
Mac.  /tmp and open() are not available inside this sandbox.

We work around this by trying open() first and by falling back to
shm_open().  On the Mac, shm_open() is a special syscall which works
inside the sandbox.  (In contrast, on Linux, shm_open() is implemented
in terms of open().)

This means that dynamic loading will work in standalone NaCl or with
--no-sandbox.  We will need to change this to use RPC to get SHM FDs.
The Chromium code for RPC in the sandbox is currently Linux-specific.

BUG=http://code.google.com/p/nativeclient/issues/detail?id=467

Review URL: http://codereview.chromium.org/1810007

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@2160 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2