msg cleanup redux.  same code as before, but in addition we replace all return statements in main with a direct call to exit.  this is because OSX's copy of SDL does not pass the return value of main (renamed to SDL_main) out through SDL's replacment main, so regardless of main's return value, OSX sel_ldr always exits with 0.
NB: the old code used to LOG_FATAL w/o giving higher-level code an
opportunity to print out an informative error message, and NaClLog on
LOG_FATAL invokes abort(3).  this should have terminated the process
via SIGABRT, which should result in an exit status of 6 in the wait
status, which bash should convert to 134 or -6, depending on whether
characters are signed or not.  i'm not sure why the old code worked w/
our test infrastructure, since the test expected an exit status of 1.

this is To Be Investigated.


Review URL: http://codereview.chromium.org/1826001

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@2212 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2