ARM validator: Add tests for treatment of branches
Add tests for branches inside and outside the chunk of code being
validated.

In order to make these tests more realistic, I have changed the test
executables to be built with the normal start address for untrusted
code (0x20000), instead of the trusted code's start address
(0x70000000).  This allows us to test for branches which underflow to
target top-bit-set addresses.  Hence the *.err golden files have
changed.

This change is in preparation for changing the ARM validator to be
more permissive by allowing bundle-aligned jumps outside the chunk of
code being validated.  Since adding tests was non-trivial, this is a
separate change.

For convenience, add the tests to validator_tests (which is faster to
run than small_tests).

BUG=http://code.google.com/p/nativeclient/issues/detail?id=481
TEST=validator_tests

Review URL: http://codereview.chromium.org/1957006

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@2215 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2