Allow NACL_DANGEROUS_* options to work when sel_ldr is run directly
The recently-added env vars only worked in the case where the
executable is loaded over SRPC, i.e. from the browser plugin.  Change
them to work when the executable is specified as a command line
argument, which is a less risky use case.

For testing nacl-glibc from the Scons build, we need a way of enabling
file access without ignoring validator results.

Update the comment because this is not entirely Chromium-specific, and
the non-browser use case at least is not intended to be temporary.

BUG=http://code.google.com/p/nativeclient/issues/detail?id=594

Review URL: http://codereview.chromium.org/2871017

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@2551 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2