Plugin: Unify plugin_unittest.cc and npnacl_unittest.cc
npnacl_unittest.cc was unused, so remove it.

npnacl_unittest.cc was mostly the same as plugin_unittest.cc, with the
exception of a test for NP_GetEntryPoints, which the plugin only
defines on Windows and Mac.  I have copied this test into
plugin_unittest.cc and added a conditional.

Also move plugin_unittest.cc into the "npapi" subdir since it is
NPAPI-specific.

TEST=run_plugin_unittest
BUG=none

Review URL: http://codereview.chromium.org/2832026

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@2579 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2