Some cleanup on the directory structure and tool naming.
The common practice with toolchains is to make the different components share an
install dir. We already did that with binutils and llvm-gcc, this change moves
llvm and llvm-gcc (with binutils) together. llvm-gcc doesn't call the llvm
tools, but we call both and tracking a single install dir makes the scripts
a bit simpler.

The other convention is that tools that are arch specific start with <arch>-.
This was true for all the tools, except for llvm-gcc.

The llvm-fake-* links stay unchanged as they are (modulo incomplete
implementations) architecture independent.

Review URL: http://codereview.chromium.org/3039022

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@2772 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2