Modular build script: Hash file contents instead of mtimes/inode numbers
This should prevent unneeded rebuilds when files are re-generated
without change, especially when intermediate trees are re-hardlinked.
Previously we were including directory inodes/mtimes, which would
always change if a hardlinked intermediate tree is re-generated.

Although this is slower, it only happens after a target is built, so
we would expect the target's output tree to be in the cache already.

BUG=http://code.google.com/p/nativeclient/issues/detail?id=484
TEST=python test_all.py

Review URL: http://codereview.chromium.org/3064003

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@2774 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2