Fix plugin to build correctly on Mac OS X 10.6
The "-bundle" flag was not being added when linking the plugin, which
leads to it being linked as an executable which cannot be dlopen()'d.

This was being caused by use of FilterOut() to remove "-bundle" on a
Clone()'d Scons env.  For some reason, this was affecting the original
"env" object's settings.  This might be a bug in FilterOut().  It
might have been caused by switching to Python 2.6, which OS 10.6 uses.
(OS X 10.5 uses Python 2.5.)

We fix this by removing those uses of FilterOut(), because this makes
the Scons rules clearer anyway.

BUG=http://code.google.com/p/nativeclient/issues/detail?id=713
TEST=run_plugin_unittest, run_fake_browser_test

Review URL: http://codereview.chromium.org/2811070

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@2800 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2