IMC: Enable test for shared memory segments
Before this, we had no tests covering imc_mem_obj_create().

I want to enable this test to be sure that imc_mem_obj_create() does
not break in the Mac OS X sandbox.

Remove the "check_contents" parameter because we might as well always
check that the shared memory region is zero-initialised.

NaCl doesn't support MAP_PRIVATE on SHM FDs, so remove the tests that
expect it to work.  Add a test that checks that MAP_PRIVATE fails.
However, we cannot enable it yet because MAP_PRIVATE causes a fatal
error.

BUG=http://code.google.com/p/nativeclient/issues/detail?id=127
BUG=http://code.google.com/p/nativeclient/issues/detail?id=724
TEST=run_imc_shm_mmap_test

Review URL: http://codereview.chromium.org/3063010

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@2803 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2