NPAPI Plugin clean-up and fixes:
1) Replace errouneous cast to ScriptableHandle with a cast to NPObject
in UrlAsNaClDesc() and pass that to UrlAsNaClDescNotify() instead of void*
that needs to be reconverted.

2) Rename and move GetObject() to npapi_native, where all the other
to/from NPVariant helpers are. Remove duplicate code by changing
it to use NPVariantToScalar.

3) Do not return false for NACL_SRPC_ARG_TYPE_OBJECT when preallocating
array-type output params in MarshallInputs(). Output parameters of such
type are prefectly legal and are in fact returned by various predeclared
PLUGIN methods. The only reason why we have caught this is because
those methods are errouneously registered as returning handles. This will
be corrected in a different CL.

4) Rename and refactor duplicate code in OutputOneDesriptor().

5) Add a comment and an assert for NACL_SRPC_ARG_TYPE_OBJECT in 
MarshallOutputs() to clarify that such objects are only returned
from internal plugin methods and therefore known to represent ScriptableHadle.

Review URL: http://codereview.chromium.org/3144009

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@3007 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2