Copy and simplify the small bits of crtbegin.o and crtend.o that we need. Also enable crtend.o.
Rationale:

The gcc/libgcc build process was created to be able to build libgcc
and therefore has to handle inline asm. To be able to do that,
it cannot use the arm driver/cc1 since they don't know anything
about intel registers for example.

The problem is that without modifying gcc the x86-64 cc1
will produce code that assumes 64 bit pointer for example.
This was not an issue for the bits of libgcc that we use
but is for crtend.o.

Fortunately these files don't depend on assembly and
can be compiled with the ARM cc1 and driver.

Instead of hacking even more the the gcc build process,
just copy the bits we need.

This is also a very small first step in a transition to
compiler-rt, libunwind and clang.

BUG=http://code.google.com/p/nativeclient/issues/detail?id=806

Review URL: http://codereview.chromium.org/3325008

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@3198 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2